修改价格float类型为BigDecimal

+++ src/main/java/com/connext/wms/entity/OutRepoItem.java
@@ -2,6 +2,8 @@ package com.connext.wms.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @author xiamingxing
  * @date 2019/1/19 3:30
@@ -10,10 +12,10 @@ import lombok.Data;
 public class OutRepoItem {
     private String goodsSku;
     private String goodsName;
-    private float price;
+    private BigDecimal price;
     private Integer num;
 
-    public OutRepoItem(String goodsSku, String goodsName, float price, Integer num) {
+    public OutRepoItem(String goodsSku, String goodsName, BigDecimal price, Integer num) {
         this.goodsSku = goodsSku;
         this.goodsName = goodsName;
         this.price = price;
